frontend: move syncdone and statuschanged to accountsync
Internally syncdone and statuschanged use the legacy events, but
from api consumer point it doesn't matter. Moved to accountsync
as that makes more sense.
