+++ frontends/web/src/routes/account/account.tsx
@@ -20,12 +20,11 @@ import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
 import { useLoad } from '../../hooks/api';
 import * as accountApi from '../../api/account';
-import { syncAddressesCount } from '../../api/accountsync';
+import { syncAddressesCount, statusChanged, syncdone } from '../../api/accountsync';
 import { TDevices } from '../../api/devices';
 import { getExchangeBuySupported, SupportedExchanges } from '../../api/exchanges';
 import { useSDCard } from '../../hooks/sdcard';
 import { unsubscribe, UnsubscribeList } from '../../utils/subscriptions';
-import { statusChanged, syncdone } from '../../api/subscribe-legacy';
 import { alertUser } from '../../components/alert/Alert';
 import { Balance } from '../../components/balance/balance';
 import { AccountGuide } from './guide';
